Refactor and organizate funcionalities and upgrade other features. Create TaskForm and update CardTask
