+++ app/Http/Controllers/HomeController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+use Inertia\Inertia;
+
+
+class HomeController extends Controller
+{
+    public function index()
+    {
+        
+    }
+}
