+++ package.json
@@ -23,6 +23,7 @@
         "@mui/icons-material": "^6.1.7",
         "@mui/material": "^6.1.7",
         "lodash.debounce": "^4.0.8",
+        "moment": "^2.30.1",
         "react": "^18.3.1",
         "react-dom": "^18.3.1"
     }
