+++ resources/css/app.css
@@ -6,6 +6,7 @@
 @tailwind components;
 @tailwind utilities;
 
+
 .btn-warning{
     color: yellow;
 }
