--- tests/Feature/ExampleTest.php
@@ -1,7 +1,0 @@
-<?php
-
-it('returns a successful response', function () {
-    $response = $this->get('/');
-
-    $response->assertStatus(200);
-});
